+++ Games/Galaga/main.cpp
@@ -530,8 +530,8 @@ private:
         }
         divers.clear();
 
-        particles.burst(playerX * tileSize + PLAYER_W * tileSize / 2,
-                        (PLAYER_Y + 1) * tileSize, 26, {120, 235, 255, 255},
+        particles.burst((float)(playerX * tileSize + PLAYER_W * tileSize / 2),
+                        (float)((PLAYER_Y + 1) * tileSize), 26, {120, 235, 255, 255},
                         10.0f, 0.7f, 6.0f);
         if (lives <= 0) {
             bestScore = std::max(bestScore, score);
